refactor: TodoCellContent 파일 분리
