+++ ProjectManager/ProjectManager/Presentation/TodoList/ViewModel/TodoListViewModel.swift
@@ -109,25 +109,3 @@ final class DefaultTodoListViewModel: TodoListViewModel {
         }
     }
 }
-
-struct TodoCellContent {
-    let title: String?
-    let body: String?
-    let deadlineAt: String
-    let id: UUID
-    let isPast: Bool
-    
-    private let dateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy. MM. d"
-        return dateFormatter
-    }()
-    
-    init(entity: TodoModel, isPast: Bool) {
-        self.title = entity.title
-        self.body = entity.body
-        self.deadlineAt = entity.deadlineAt.toString(dateFormatter)
-        self.id = entity.id
-        self.isPast = isPast
-    }
-}
